Don't compile systray with mobile or JS
